wip: create blk, vcpu, types files
