+++ type.h
@@ -1,0 +1,6 @@
+#include <stdint.h>
+
+typedef uint8_t u8;
+typedef uint8_t u16;
+typedef uint32_t u32;
+typedef uint64_t u64;
